FTAS: move SafePipe to utils
This way we can use SafePipe elsewhere and not just in FTAS.
